requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Microsoft GraphRAG
graphrag

# LLM and Embeddings
openai

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
langchain-text-splitters

# Data processing
pandas
numpy
pyarrow
